+++ NEWS.md
@@ -14,6 +14,10 @@ New features and enhancements:
   to more easily label spikes. (#203; thanks @mattansb for the suggestion).
 * The `arrow` parameter is now supported for intervals in `geom_slabinterval()`
   (#206; thanks to @ASKurz for the suggestion).
+* The color ramp scales (e.g. `scale_colour_ramp_continuous()`, ...) now use
+  an explicit data type, `partial_colour_ramp()`, to encode color ramps and 
+  their origin colors. This should make it easier to apply explicit color ramps
+  without using scale functions, if needed (#209).
 
 Documentation: